+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/MsgFormatTest.groovy
@@ -14,22 +14,28 @@ class MsgFormatTest extends Specification {
     def "Check format to file"() {
 
         when: "format"
-        String res = FormatUtils.formatForFile(new ReportMessage(source: "some.Sample.java", line: 2, code: 'bla'))
+        def message = new ReportMessage(source: "some.Sample.java", line: 2, code: 'bla', signature: 'java16-sun-1.0')
+        String res = FormatUtils.formatForFile(message, false)
         then: "correct"
         res == 'some.Sample:2  Undefined reference: bla'
 
+        when: "format with signature"
+        res = FormatUtils.formatForFile(message, true)
+        then: "correct"
+        res == 'some.Sample:2  Undefined reference (java16-sun-1.0): bla'
+
         when: "no line"
-        res = FormatUtils.formatForFile(new ReportMessage(source: "some.Sample.java", code: 'bla'))
+        res = FormatUtils.formatForFile(new ReportMessage(source: "some.Sample.java", code: 'bla'), false)
         then: "correct"
         res == 'some.Sample  Undefined reference: bla'
 
         when: "incorrect source"
-        res = FormatUtils.formatForFile(new ReportMessage(source: "failed source", code: 'bla'))
+        res = FormatUtils.formatForFile(new ReportMessage(source: "failed source", code: 'bla'), false)
         then: "correct"
         res == 'failed source  Undefined reference: bla'
 
         when: "failed parse"
-        res = FormatUtils.formatForFile(new ReportMessage(parseFail: true, code: 'bla bla'))
+        res = FormatUtils.formatForFile(new ReportMessage(parseFail: true, code: 'bla bla'), false)
         then: "correct"
         res == 'bla bla'
     }
@@ -37,22 +43,28 @@ class MsgFormatTest extends Specification {
     def "Check format for console"() {
 
         when: "format"
-        String res = FormatUtils.formatForConsole(new ReportMessage(source: "some.Sample.java", line: 2, code: 'bla'))
+        def message = new ReportMessage(source: "some.Sample.java", line: 2, code: 'bla', signature: 'java16-sun-1.0')
+        String res = FormatUtils.formatForConsole(message, false)
         then: "correct"
         res.replaceAll('\r', '') == '[Undefined reference] some.(Sample.java:2)\n  >> bla\n'
 
+        when: "format with signature"
+        res = FormatUtils.formatForConsole(message, true)
+        then: "correct"
+        res.replaceAll('\r', '') == '[Undefined reference (java16-sun-1.0)] some.(Sample.java:2)\n  >> bla\n'
+
         when: "no line"
-        res = FormatUtils.formatForConsole(new ReportMessage(source: "some.Sample.java", code: 'bla'))
+        res = FormatUtils.formatForConsole(new ReportMessage(source: "some.Sample.java", code: 'bla'), false)
         then: "correct"
         res.replaceAll('\r', '') == '[Undefined reference] some.(Sample.java:1)\n  >> bla\n'
 
         when: "incorrect source"
-        res = FormatUtils.formatForConsole(new ReportMessage(source: "failed source", code: 'bla'))
+        res = FormatUtils.formatForConsole(new ReportMessage(source: "failed source", code: 'bla'), false)
         then: "correct"
         res.replaceAll('\r', '') == '[Undefined reference] failed source\n  >> bla\n'
 
         when: "failed parse"
-        res = FormatUtils.formatForConsole(new ReportMessage(parseFail: true, code: 'bla bla'))
+        res = FormatUtils.formatForConsole(new ReportMessage(parseFail: true, code: 'bla bla'), false)
         then: "correct"
         res.replaceAll('\r', '') == '[Unrecognized error] bla bla \n'
     }