+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/resources/ResourceCasesKitTest.groovy
@@ -26,7 +26,7 @@ class ResourceCasesKitTest extends AbstractKitTest {
             repositories { mavenCentral()}
             dependencies {
                 signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
 
         """
@@ -74,8 +74,8 @@ class ResourceCasesKitTest extends AbstractKitTest {
             repositories { mavenCentral() }
             dependencies {
                 signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-                signature 'net.sf.androidscents.signature:android-api-level-14:4.0_r4'
-                compile 'junit:junit:4.12'
+                signature 'net.sf.androidscents.signature:android-api-level-14:4.0_r4@signature'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
         """
         fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
@@ -115,7 +115,7 @@ class ResourceCasesKitTest extends AbstractKitTest {
 
             repositories { mavenCentral() }
             dependencies {
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
         """
         fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
@@ -151,7 +151,7 @@ class ResourceCasesKitTest extends AbstractKitTest {
             repositories { mavenCentral() }
             dependencies {
                 signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
         """
         fileFromClasspath('src/main/java/invalid/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java')
@@ -186,7 +186,7 @@ class ResourceCasesKitTest extends AbstractKitTest {
             repositories { mavenCentral()}
             dependencies {
                 signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
 
         """
@@ -222,7 +222,7 @@ class ResourceCasesKitTest extends AbstractKitTest {
             repositories { mavenCentral()}
             dependencies {
                 signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
 
         """