+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/signature/BuildSignatureExcludeKitTest.groovy
@@ -27,7 +27,7 @@ class BuildSignatureExcludeKitTest extends AbstractKitTest {
             repositories { mavenCentral()}
             dependencies {
                 compile "org.codehaus.mojo:animal-sniffer-annotations:1.14"
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
         """
         fileFromClasspath('src/main/java/ann/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/ann/Sample.java')
@@ -62,7 +62,7 @@ class BuildSignatureExcludeKitTest extends AbstractKitTest {
             repositories { mavenCentral()}
             dependencies {
                 compile "org.codehaus.mojo:animal-sniffer-annotations:1.14"
-                compile 'junit:junit:4.12'
+                compile 'org.slf4j:slf4j-api:1.7.25'
             }
         """
         fileFromClasspath('src/main/java/ann/Sample.java', '/ru/vyarus/gradle/plugin/animalsniffer/java/ann/Sample.java')