+++ src/test/resources/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample.java
@@ -1,6 +1,6 @@
 package invalid;
 
-import org.junit.Assert;
+import org.slf4j.LoggerFactory;
 
 import java.nio.file.Paths;
 
@@ -14,6 +14,6 @@ public class Sample {
     public void someth() {
         // class added in 1.7
         Paths.get("/tmp");
-        Assert.assertTrue(true);
+        LoggerFactory.getLogger(Sample.class);
     }
 }