+++ src/test/resources/ru/vyarus/gradle/plugin/animalsniffer/java/invalid/Sample2.java
@@ -1,6 +1,6 @@
 package invalid;
 
-import org.junit.Assert;
+import org.slf4j.LoggerFactory;
 
 import java.nio.file.FileSystems;
 
@@ -9,6 +9,6 @@ public class Sample2 {
     public static void main(String[] args) {
         // added in 1.7
         FileSystems.getDefault().getFileStores();
-        Assert.assertTrue(true);
+        LoggerFactory.getLogger(Sample2.class);
     }
 }