+++ src/test/resources/ru/vyarus/gradle/plugin/animalsniffer/java/valid/Sample_1.java
@@ -1,6 +1,6 @@
 package valid;
 
-import org.junit.Assert;
+import org.slf4j.LoggerFactory;
 
 public class Sample {
 
@@ -11,7 +11,7 @@ public class Sample {
     }
 
     public static void main(String[] args) {
-        Assert.assertTrue(true);
+        LoggerFactory.getLogger(Sample.class);
     }
 
     // new method to change signature and avoid gradle smart detectors